# cmd/swagger/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.DeleteRoundReq:
    properties:
      groupID:
        type: string
      roundID:
        type: string
      version:
        type: integer
    type: object
...
  dto.GameActivity:
    properties:
      count:
        type: integer
      last_date:
        type: string
    type: object
...
  dto.Page:
    properties:
      activity:
        additionalProperties:
          $ref: '#/definitions/dto.GameActivity'
        type: object
      games:
        type: integer
      items:
        items:
          $ref: '#/definitions/dto.Round'
        type: array
      players:
        type: integer
      stats:
        items:
          $ref: '#/definitions/dto.Stat'
        type: array
      total:
        type: integer
    type: object
...
  dto.Recap:
    properties:
      from:
        type: string
      games:
        type: integer
      minutes:
        type: integer
      period:
        type: string
      photos:
        items:
          type: string
        type: array
      players:
        type: integer
      rounds:
        type: integer
      to:
        type: string
      top_game:
        type: string
      top_game_rounds:
        type: integer
      top_player:
        type: string
      top_plays:
        type: integer
    type: object
  dto.RestoreRoundReq:
    properties:
      groupID:
        type: string
      roundID:
        type: string
      version:
        type: integer
    type: object
  dto.Round:
    properties:
      author:
        type: string
      date:
        type: string
      deleted_at:
        type: string
      game_id:
        type: string
      id:
        type: string
      location:
        type: string
      memory:
        type: string
      minutes:
        type: integer
      mode:
        type: string
      outcome:
        type: string
      photos:
        items:
          type: string
        type: array
      players:
        items:
          type: string
        type: array
      scores:
        additionalProperties:
          type: string
        type: object
      team_score:
        type: string
      teams:
        items:
          $ref: '#/definitions/dto.Team'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
      winners:
        items:
          type: string
        type: array
    type: object
...
  dto.Stat:
    properties:
      game:
        type: string
      mode:
        type: string
      played:
        type: integer
      player:
        type: string
      samples:
        type: integer
      wins:
        type: integer
    type: object
  dto.Team:
    properties:
      id:
        type: string
      name:
        type: string
      players:
        items:
          type: string
        type: array
      score:
        type: string
      winner:
        type: boolean
    type: object
...
  ginutils.Ret-any:
    properties:
      code:
        type: integer
      data: {}
      message: {}
    type: object
...
  ginutils.Ret-array_dto_Round:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.Round'
        type: array
      message: {}
    type: object
...
  ginutils.Ret-dto_Page:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Page'
      message: {}
    type: object
...
  ginutils.Ret-dto_Recap:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Recap'
      message: {}
    type: object
  ginutils.Ret-dto_Round:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.Round'
      message: {}
    type: object
...
info:
  contact: {}
  description: |-
    熟人小组对局日记的接口文档；路径相对 /api 挂载点。
    除图片读取外，所有接口返回 ginutils.Ret 包络：{code, data, message}；code 为稳定业务码，失败时 HTTP 状态同时表达语义。
  title: 又一局 API
  version: "1.0"
paths:
...
  /v1/groups/{group}/rounds:
    get:
      description: 按日期、桌游、玩家筛选对局，返回分页结果与战绩统计
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: 开始日期
        in: query
        name: from
        type: string
      - description: 结束日期
        in: query
        name: to
        type: string
      - description: 桌游 ID
        in: query
        name: game
        type: string
      - description: 玩家 ID
        in: query
        name: player
        type: string
      - description: 搜索回忆或地点
        in: query
        name: q
        type: string
      - description: 地点
        in: query
        name: location
        type: string
      - description: individual/team/coop
        in: query
        name: mode
        type: string
      - description: win/loss/draw/unknown
        in: query
        name: outcome
        type: string
      - description: 是否有照片
        in: query
        name: has_photos
        type: boolean
      - description: 偏移量
        in: query
        name: offset
        type: integer
      - description: 每页条数，最大 100
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Page'
      security:
      - SessionCookie: []
      summary: 获取对局列表
      tags:
      - Round
    post:
      consumes:
      - application/json
      description: 创建对局需携带 Idempotency-Key；修改对局需携带版本号
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: 对局 ID，修改时必填
        in: path
        name: id
        type: string
      - description: 提交标识，创建时必填
        in: header
        name: Idempotency-Key
        type: string
      - description: 对局内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Round'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Round'
      security:
      - SessionCookie: []
      summary: 记一局或修改对局
      tags:
      - Round
  /v1/groups/{group}/rounds/{id}:
    delete:
      consumes:
      - application/json
      description: 按版本号移入七天回收站，版本不一致时提示刷新
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: 对局 ID
        in: path
        name: id
        required: true
        type: string
      - description: 删除对局请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteRoundReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-any'
      security:
      - SessionCookie: []
      summary: 删除对局
      tags:
      - Round
    get:
      description: 返回一条对局的完整内容，含版本号
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: 对局 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Round'
      security:
      - SessionCookie: []
      summary: 获取对局详情
      tags:
      - Round
  /v1/groups/{group}/rounds/{id}/restore:
    post:
      consumes:
      - application/json
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: 对局 ID
        in: path
        name: id
        required: true
        type: string
      - description: 恢复请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RestoreRoundReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Round'
      security:
      - SessionCookie: []
      summary: 恢复已删除对局
      tags:
      - Round
  /v1/groups/{group}/rounds/recap:
    get:
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      - description: YYYY-MM 或 YYYY
        in: query
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-dto_Recap'
      security:
      - SessionCookie: []
      summary: 获取月度或年度回顾
      tags:
      - Round
  /v1/groups/{group}/rounds/recycle-bin:
    get:
      parameters:
      - description: 小组 ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ginutils.Ret-array_dto_Round'
      security:
      - SessionCookie: []
      summary: 获取回收站
      tags:
      - Round
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: omr_session
    type: apiKey
swagger: "2.0"